Recommender

Recommender.AbstractDataset

Recommender.Dataset

Recommender.kNNRecommender

Base.download

MLJModelInterface.fit

Recommender.dataset

Recommender.datasets

Recommender.download_zip

Recommender.hitrate

Recommender.load

Recommender.AbstractDataset — Type

AbstractDataset

Abstract type for dataset objects.

Recommender.Dataset — Type

Dataset

Dataset struct to manage concrete datasets.

Recommender.kNNRecommender — Type

kNNRecommender

k: compute k nearest neighbor similarity.

shrink: if nonzero, decrease contributions from items with few rating.

weighting: currently supoorts TF-IDF

npred: number of retrieval by predict.

Base.download — Method

download(dataset::Dataset; <keyword arguments>)

Download actual dataset to /dataset/.

Arguments

Keyword arguments depend on dataset format.

dataset.format=zip: see download_zip.

MLJModelInterface.fit — Function

fit(model::kNNRecommender, verbosity::Int, X, y, w=nothing)

fit kNNRecommender. For ItemkNN, X = userids, y = itemids. w is additional weighting, for instance, rating of user to item.

Recommender.dataset — Method

dataset(name::AbstractString)

Return Dataset object specified by name.

Recommender.datasets — Method

datasets()

Return dataframe of all the dataset information from datasets.csv.

Recommender.download_zip — Method

download_zip(url::AbstractString, name::AbstractString, dirpath::String; <keyword arguments>)

Download zipped dataset name from url, and place it to dirpath.

Keyword arguments

usecache=true: if true and dirpath already exists, it skips download. If false, it force download by removing dirpath (if exists).

unzip=true: whether extract datasets from zip.

removezip=false: whether remove original zip after extraction.

Recommender.hitrate — Method

hitrate(recommends, ys)

Compute mean HR.

arguments

recommends: recommends[i] is list of predictions for i-th input.

ys: ys[i] is ground-truth for i-th input.

jldoctest julia> hitrate([[1, 3], nothing, [2, 4]], [1, 2, 3]) 0.3333333333333`

Recommender.load — Method

load(dataset::Dataset)

Load dataset. Return type depends on each dataset. For instance, movielens 1M is loaded as DataFrame.