Recommender

Recommender.AbstractDataset

Recommender.Movielens100k

Recommender.download_zip

Recommender.AbstractDataset — Type

AbstractDataset

Abstract type for dataset objects.

Recommender.Movielens100k — Type

Movielens100k

Recommender.download_zip — Method

download_zip(url::AbstractString, name::AbstractString, dirpath::String; <keyword arguments>)

Download zipped dataset name from url, and place it to dirpath.

Keyword arguments

usecache=true: if true and dirpath already exists, it skips download. If false, it force download by removing dirpath (if exists).

unzip=true: whether extract datasets from zip.

removezip=false: whether remove original zip after extraction.